+++ src/Resources/assets/js/contao-youtube-bundle-theme.js
@@ -1,1 +1,2 @@
 require('../scss/contao-youtube-bundle.scss');
+require('alertifyjs/build/css/alertify.min.css');