+++ src/Resources/public/js/contao-youtube-bundle-theme.js
@@ -1,1 +1,1 @@
-"use strict";require("../scss/contao-youtube-bundle.scss");
+"use strict";require("../scss/contao-youtube-bundle.scss"),require("alertifyjs/build/css/alertify.min.css");